platform: add 'depends' property handler
Add 'depends' property handler.
The purpose is to create a kind of dependency between devices.
The field contains the name of the device that the loaded
driver depends on. The device driver reads this property
and decides either continue or postpone the device probing.
This is an optional property. The fact that the device node does
not have such the property makes no changes in the device probing.

Signed-off-by: Valentin Raevsky <valentin@compulab.co.il>