ARM: i.MX6: cl-som-imx6: dts: set USB H1 power GPIO
In U2 configuration the USB hub is not installed.
When the USB is not installed the value of the USB H1 power signal is
set to 0, disabling the USB.

Set the pin MUX of the signal USBH1_PWR to GPIO.
Use a regulator to set the value of the signal USBH1_PWR to 1.

Signed-off-by: Uri Mashiach <uri.mashiach@compulab.co.il>